ADD: Add symbol field support to CSV and JSON
