+++ c/src/text_serialization.rs
@@ -179,10 +179,10 @@ pub unsafe extern "C" fn s_serialize_record(
     let res = match options.encoding {
         TextEncoding::Json => {
             json::Encoder::new(&mut cursor, false, options.pretty_px, options.pretty_ts)
-                .encode_record_ref(record, options.ts_out)
+                .encode_record_ref_ts_out(record, options.ts_out)
         }
         TextEncoding::Csv => csv::Encoder::new(&mut cursor, options.pretty_px, options.pretty_ts)
-            .encode_record_ref(record, options.ts_out),
+            .encode_record_ref_ts_out(record, options.ts_out),
     }
     // null byte
     .and_then(|_| {
@@ -232,10 +232,10 @@ pub unsafe extern "C" fn f_serialize_record(
     let res = match options.encoding {
         TextEncoding::Json => {
             json::Encoder::new(&mut cfile, false, options.pretty_px, options.pretty_ts)
-                .encode_record_ref(record, options.ts_out)
+                .encode_record_ref_ts_out(record, options.ts_out)
         }
         TextEncoding::Csv => csv::Encoder::new(&mut cfile, options.pretty_px, options.pretty_ts)
-            .encode_record_ref(record, options.ts_out),
+            .encode_record_ref_ts_out(record, options.ts_out),
     };
     if res.is_ok() {
         cfile.bytes_written() as i32
@@ -266,5 +266,5 @@ fn serialize_csv_header<W: io::Write, R: DbnEncodable>(
     _rec: &R,
     encoder: &mut csv::Encoder<W>,
 ) -> dbn::Result<()> {
-    encoder.encode_header::<R>()
+    encoder.encode_header::<R>(false)
 }
